decouple divider and scrollbar
