+++ static/app/components/events/interfaces/spans/utils.tsx
@@ -21,78 +21,6 @@ import {
   TreeDepthType,
 } from './types';
 
-type Rect = {
-  // x and y are left/top coords respectively
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-};
-
-// get position of element relative to top/left of document
-const getOffsetOfElement = (element: Element) => {
-  // left and top are relative to viewport
-  const {left, top} = element.getBoundingClientRect();
-
-  // get values that the document is currently scrolled by
-  const scrollLeft = window.pageXOffset;
-  const scrollTop = window.pageYOffset;
-
-  return {x: left + scrollLeft, y: top + scrollTop};
-};
-
-export const rectOfContent = (element: Element): Rect => {
-  const {x, y} = getOffsetOfElement(element);
-
-  // offsets for the border and any scrollbars (clientLeft and clientTop),
-  // and if the element was scrolled (scrollLeft and scrollTop)
-  //
-  // NOTE: clientLeft and clientTop does not account for any margins nor padding
-  const contentOffsetLeft = element.clientLeft - element.scrollLeft;
-  const contentOffsetTop = element.clientTop - element.scrollTop;
-
-  return {
-    x: x + contentOffsetLeft,
-    y: y + contentOffsetTop,
-    width: element.scrollWidth,
-    height: element.scrollHeight,
-  };
-};
-
-export const rectOfViewport = (): Rect => ({
-  x: window.pageXOffset,
-  y: window.pageYOffset,
-  width: window.document.documentElement.clientWidth,
-  height: window.document.documentElement.clientHeight,
-});
-
-export const rectRelativeTo = (rect: Rect, pos = {x: 0, y: 0}): Rect => ({
-  x: rect.x - pos.x,
-  y: rect.y - pos.y,
-  width: rect.width,
-  height: rect.height,
-});
-
-export const rectOfElement = (element: HTMLElement): Rect => {
-  const {x, y} = getOffsetOfElement(element);
-  return {
-    x,
-    y,
-    width: element.offsetWidth,
-    height: element.offsetHeight,
-  };
-};
-
-export const clamp = (value: number, min: number, max: number): number => {
-  if (value < min) {
-    return min;
-  }
-  if (value > max) {
-    return max;
-  }
-  return value;
-};
-
 export const isValidSpanID = (maybeSpanID: any) =>
   isString(maybeSpanID) && maybeSpanID.length > 0;
